# app/lib/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "type": "object",
        "description": "Password change request.",
        "required": ["old_password", "new_password"],
        "properties": {
          "new_password": {
            "type": "string",
            "description": "Replacement password."
          },
          "old_password": {
            "type": "string",
            "description": "Current password."
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "description": "FastAPI-compatible error envelope.",
        "required": ["detail", "code", "retryable"],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error category."
          },
          "detail": {
            "type": "string",
            "description": "Error detail payload."
          },
          "retryable": {
            "type": "boolean",
            "description": "Whether retrying the request may succeed without user correction."
          }
        }
      },
...
      "InviteCodeResponse": {
        "type": "object",
        "description": "Invite code response.",
        "required": [
          "id",
          "code",
          "used",
          "status",
          "expires_at",
          "max_uses",
          "use_count",
          "created_at"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Raw invite code."
          },
          "created_at": {
            "type": "number",
            "format": "double",
            "description": "Invite code creation timestamp."
          },
          "expires_at": {
            "type": "number",
            "format": "double",
            "description": "Absolute expiration timestamp."
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Invite code row identifier."
          },
          "max_uses": {
            "type": "integer",
            "format": "int64",
            "description": "Maximum permitted registrations."
          },
          "revoked_at": {
            "type": ["number", "null"],
            "format": "double",
            "description": "Optional irreversible revocation timestamp."
          },
          "status": {
            "$ref": "#/components/schemas/InviteCodeStatus",
            "description": "Current lifecycle state."
          },
          "use_count": {
            "type": "integer",
            "format": "int64",
            "description": "Number of committed registrations."
          },
          "used": {
            "type": "boolean",
            "description": "Whether the invite code has been consumed."
          }
        }
      },
      "InviteCodeStatus": {
        "type": "string",
        "description": "Public lifecycle state for an invite code.",
        "enum": ["active", "expired", "revoked", "exhausted"]
      },
...
      "LoginRequest": {
        "type": "object",
        "description": "Login request.",
        "required": ["username", "password"],
        "properties": {
          "password": {
            "type": "string",
            "description": "Password."
          },
          "username": {
            "type": "string",
            "description": "Username."
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "description": "Login response that intentionally omits the raw session token.",
        "required": ["user", "expires_at"],
        "properties": {
          "expires_at": {
            "type": "number",
            "format": "double",
            "description": "Session expiration timestamp."
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse",
            "description": "Authenticated user."
          }
        }
      },
      "LogoutResponse": {
        "type": "object",
        "description": "Logout response.",
        "required": ["ok", "user_id"],
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "Whether the operation succeeded."
          },
          "user_id": {
            "$ref": "#/components/schemas/UserId",
            "description": "Authenticated user identifier."
          }
        }
      },
...
      "OkResponse": {
        "type": "object",
        "description": "Boolean ok response.",
        "required": ["ok"],
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "Whether the operation succeeded."
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "description": "Account registration request.",
        "required": ["username", "password", "invite_code"],
        "properties": {
          "invite_code": {
            "type": "string",
            "description": "Invite code text required for every public registration."
          },
          "password": {
            "type": "string",
            "description": "Requested password."
          },
          "username": {
            "type": "string",
            "description": "Requested username."
          }
        }
      },
...
      "SessionRevocationErrorDetail": {
        "type": "object",
        "description": "Stable detail returned when durable session revocation cannot be confirmed.",
        "required": ["code", "message", "request_id"],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable client classification."
          },
          "message": {
            "type": "string",
            "description": "Safe user-facing failure summary."
          },
          "request_id": {
            "type": "string",
            "description": "Server-generated request identifier for audit correlation."
          }
        }
      },
      "SessionRevocationErrorResponse": {
        "type": "object",
        "description": "Error envelope for an unconfirmed session revocation.",
        "required": ["detail"],
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/SessionRevocationErrorDetail",
            "description": "Structured revocation failure detail."
          }
        }
      },
      "TokenCreateRequest": {
        "type": "object",
        "description": "Access token creation request.",
        "properties": {
          "name": {
            "type": "string",
            "description": "Token display name.",
            "maxLength": 100
          },
          "ttl": {
            "type": "integer",
            "format": "int64",
            "description": "Requested token TTL in seconds.",
            "maximum": 31536000,
            "minimum": 3600
          }
        }
      },
      "TokenCreateResponse": {
        "type": "object",
        "description": "Access token creation response.",
        "required": ["id", "token", "name", "expires_at"],
        "properties": {
          "expires_at": {
            "type": "number",
            "format": "double",
            "description": "Token expiration timestamp."
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Token row identifier."
          },
          "name": {
            "type": "string",
            "description": "Token display name."
          },
          "token": {
            "type": "string",
            "description": "Raw token value returned only at creation time."
          }
        }
      },
      "TokenInfo": {
        "type": "object",
        "description": "Access token metadata response.",
        "required": ["id", "name", "expires_at", "created_at"],
        "properties": {
          "created_at": {
            "type": "number",
            "format": "double",
            "description": "Token creation timestamp."
          },
          "expires_at": {
            "type": "number",
            "format": "double",
            "description": "Token expiration timestamp."
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Token row identifier."
          },
          "name": {
            "type": "string",
            "description": "Token display name."
          }
        }
      },
...
      "UserId": {
        "type": "integer",
        "format": "int64",
        "description": "Auth user identifier stored as an ordinary SQLite integer."
      },
      "UserResponse": {
        "type": "object",
        "description": "User profile returned by auth endpoints.",
        "required": ["id", "username", "is_admin"],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UserId",
            "description": "User identifier."
          },
          "is_admin": {
            "type": "boolean",
            "description": "Whether the user has admin privileges."
          },
          "username": {
            "type": "string",
            "description": "Login username."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "litradar_session"
      }
    }
  },
  "info": {
    "title": "LitRadar API",
    "description": "Generated OpenAPI document for the Rust API.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/change-password": {
      "post": {
        "tags": ["auth"],
        "summary": "Change the current user's password.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `headers` - Request headers.\n* `body` - Password change request.\n\n# Returns\n\nOK response.",
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Password changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/auth/invite-code": {
      "delete": {
        "tags": ["auth"],
        "summary": "Revoke the current user's unrevoked invite code.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `headers` - Request headers.\n* `request_id` - Server-generated request identifier.\n\n# Returns\n\nOK response when an invite was revoked.",
        "operationId": "revoke_invite_code",
        "responses": {
          "200": {
            "description": "Invite code revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "404": {
            "description": "No unrevoked invite code exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "get": {
        "tags": ["auth"],
        "summary": "Get the invite code generated by the current user.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `headers` - Request headers.\n\n# Returns\n\nInvite code response or null.",
        "operationId": "get_invite_code",
        "responses": {
          "200": {
            "description": "Current user's invite code.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/InviteCodeResponse"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "post": {
        "tags": ["auth"],
        "summary": "Generate a one-time invite code.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `headers` - Request headers.\n\n# Returns\n\nInvite code response.",
        "operationId": "generate_invite_code",
        "responses": {
          "200": {
            "description": "Generated invite code.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteCodeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/auth/invite-code/rotate": {
      "post": {
        "tags": ["auth"],
        "summary": "Revoke any prior issuance and create one replacement invite code.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `headers` - Request headers.\n* `request_id` - Server-generated request identifier.\n\n# Returns\n\nNewly issued replacement invite code.",
        "operationId": "rotate_invite_code",
        "responses": {
          "200": {
            "description": "Invite code rotated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteCodeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Authenticate a user and set a session cookie.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `body` - Login request.\n\n# Returns\n\nLogin response without the raw token.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "429": {
            "description": "Authentication rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "Logout the current session token and clear the browser cookie.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `headers` - Request headers.\n\n# Returns\n\nLogout response.",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Logged out.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "503": {
            "description": "The browser cookie was cleared but durable token revocation was not confirmed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRevocationErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/auth/logout-all": {
      "post": {
        "tags": ["auth"],
        "summary": "Revoke every session and personal access token for the current user.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `headers` - Request headers.\n* `request_id` - Server-generated request identifier.\n\n# Returns\n\nLogout response after every token is durably revoked.",
        "operationId": "logout_all",
        "responses": {
          "200": {
            "description": "All sessions and personal access tokens were revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "503": {
            "description": "The browser cookie was cleared but durable token revocation was not confirmed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRevocationErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["auth"],
        "summary": "Return the current authenticated user.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `headers` - Request headers.\n\n# Returns\n\nCurrent user response.",
        "operationId": "get_me",
        "responses": {
          "200": {
            "description": "Current user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register a new user account.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `body` - Registration request.\n\n# Returns\n\nCreated user response.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Registered user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "429": {
            "description": "Authentication rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/tokens": {
      "get": {
        "tags": ["auth"],
        "summary": "List active access tokens for the current user.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `headers` - Request headers.\n\n# Returns\n\nActive token metadata.",
        "operationId": "get_tokens",
        "responses": {
          "200": {
            "description": "Active access tokens.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TokenInfo"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "post": {
        "tags": ["auth"],
        "summary": "Create an access token for the current user.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `headers` - Request headers.\n* `body` - Token creation request.\n\n# Returns\n\nCreated token response.",
        "operationId": "create_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation order: authentication, raw name length, normalized reserved name, TTL, then quota. Errors: Access token name must be at most 100 Unicode code points; Access token name \"login\" is reserved; Access token TTL must be between 3600 and 31536000 seconds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Validation order: authentication, raw name length, normalized reserved name, TTL, then quota. Error: Active access token limit of 50 reached; revoke a token before creating another.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/auth/tokens/{token_id}": {
      "delete": {
        "tags": ["auth"],
        "summary": "Delete an access token by row id.",
        "description": "# Arguments\n\n* `state` - Shared API state.\n* `headers` - Request headers.\n* `token_id` - Token row identifier.\n\n# Returns\n\nOK response.",
        "operationId": "delete_token",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "description": "Token row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
...
}